refactor(bookings): merge requests page in and regroup navigation
Fold the /requests page into /bookings as a two-column master-detail
view: the filterable overview stays in the left column while the right
column shows the selected booking's conversation, status actions and
message thread. The conversation pane is extracted into
BookingConversationPanel so the page itself stays a list.

Selection is driven by the URL (/bookings/:bookingId?), keeping deep
links working and avoiding an effect to mirror the param into state.
Below `md` the two columns swap in place, with a back button returning
to the list.

Unread-message handling comes along with the move: rows carry the unread
badge and the header badge now sits on Bookings, so the websocket
new-message notifications surface where the conversations live. The
/requests route and its now-unused translations are removed, which also
clears four duplicate translation keys that tripped TS1117.

Also lift Items, Bookings and Collections out of the avatar dropdown
into the main navigation bar, drop the Requests button, and shorten
"+ Share item" to "+ Share". The dropdown keeps Profile and Sign Out.
